Nueva actualizacion de modulo Soluciones Construidas.

+++ habitathgrp/src/org/habitatguate/hgerp/seguridad/client/auxjdo/AuxSolicitudGarantiaHipotecaria.java
@@ -90,13 +90,13 @@ public class AuxSolicitudGarantiaHipotecaria implements IsSerializable {
 	}
 
 	
-    private int areaTerreno;
+    private float areaTerreno;
 	
-	public int getAreaTerreno() {
+	public float getAreaTerreno() {
 		return areaTerreno;
 	}
 
-	public void setAreaTerreno(int areaTerreno) {
+	public void setAreaTerreno(float areaTerreno) {
 		this.areaTerreno = areaTerreno;
 	}
